+++ src/cashier_frontend/src/components/wallet-page/header.tsx
