+++ src/cashier_frontend/src/components/wallet-page/nft-card.tsx
