+++ src/cashier_frontend/src/components/wallet-page/nfts-tab.tsx
