+++ src/cashier_frontend/src/components/wallet-page/token-card.tsx
