+++ src/cashier_frontend/src/pages/wallet/page.tsx
@@ -1,0 +1,11 @@
+import { WalletHero } from "@/components/wallet-page/hero";
+import { WalletTabs } from "@/components/wallet-page/tabs";
+
+export function WalletPage() {
+    return (
+        <>
+            <WalletHero />
+            <WalletTabs />
+        </>
+    );
+}
